--- src/pages/stripe/CheckoutForm.tsx
@@ -1,34 +1,0 @@
-import * as React from "react";
-
-const injectStripe = require("react-stripe-elements");
-
-import AddressSection from "./AddressSection";
-import CardSection from "./CardSection";
-
-class CheckoutForm extends React.Component<any, any> {
-    handleSubmit = (ev: any) => {
-        // We don't want to let default form submission happen here, which would refresh the page.
-        ev.preventDefault();
-
-        // Within the context of `Elements`, this call to createToken knows which Element to
-        // tokenize, since there's only one in this group.
-        this.props.stripe.createToken({ name: "Jenny Rosen" }).then(({ token }: any) => {
-            console.log("Received Stripe token:", token);
-        });
-
-        // However, this line of code will do the same thing:
-        // this.props.stripe.createToken({type: 'card', name: 'Jenny Rosen'});
-    }
-
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <AddressSection />
-                <CardSection />
-                <button>Confirm order</button>
-            </form>
-        );
-    }
-}
-
-export default injectStripe(CheckoutForm);
